fix references related to the rename of package-json.js file
